git-svn-id: svn+ssh://software-srn.sandia.gov/svn/phish/trunk@346 76edfe0a-4bdc-4aa6-b265-a765f6f55952

+++ minnow/rmat.py
@@ -59,6 +59,6 @@ for m in xrange(ngenerate):
   phish.pack_uint64(i)
   phish.pack_uint64(j)
   #phish.send(0)
-  phish.send_key(0,i)
+  phish.send_key(0,long(i))
 
 phish.exit()
